requirements.txt
xlwt
opencv-python
numpy
scipy
matplotlib
pillow
